Create Final_Project15_13
Here is the Project that we were able to "complete" to the best of our ability. We had three main errors that we were hoping to solve or get more guidance tomorrow in class. We had also tried to consult other students for help as well as the discussions and the book but this was what we came up with. 

Problem 1: Even though we followed Jasons solution to a T we were still unable to het Java FX to work. We found countless tutorials but most of them were outdated and not helpful. 

Problem 2: We had a small issue with declaring some of the object, pane and rectangle. It gave the error "Rectangle cannot be resolved to a type" even though the notation for creating the object was correct. Or at least we think it is. 

Problem 3: finally we just had some doubts about our BigO analysis and  wether or not its correct. But you said that any problems with that will be addressed in class so its not a super big issue.